move the recent exposures screen to its own folder
