+++ src/screens/exposureHistory/ExposureHistoryScreen.tsx
@@ -1,5 +1,5 @@
 import React, {useCallback} from 'react';
-import {StyleSheet, TouchableOpacity, Alert} from 'react-native';
+import {StyleSheet, Alert} from 'react-native';
 import {useI18n, I18n} from 'locale';
 import {CombinedExposureHistoryData, getCurrentOutbreakHistory, OutbreakHistoryItem, OutbreakSeverity} from 'shared/qr';
 import {useNavigation} from '@react-navigation/native';
@@ -7,8 +7,9 @@ import {Box, Text, Icon, Toolbar, Button} from 'components';
 import {SafeAreaView} from 'react-native-safe-area-context';
 import {ScrollView} from 'react-native-gesture-handler';
 import {useOutbreakService} from 'services/OutbreakService';
-import {formatExposedDate} from 'shared/date-fns';
 import {useExposureHistory, useClearExposedStatus} from 'services/ExposureNotificationService';
+
+import {ExposureList} from './views/ExposureList';
 
 const severityText = ({severity, i18n}: {severity: OutbreakSeverity; i18n: I18n}) => {
   switch (severity) {
@@ -53,50 +54,6 @@ const toProximityExposureHistoryData = ({
       timestamp: outbreak,
     };
   });
-};
-
-const ExposureList = ({exposureHistoryData}: {exposureHistoryData: CombinedExposureHistoryData[]}) => {
-  const i18n = useI18n();
-  const dateLocale = i18n.locale === 'fr' ? 'fr-CA' : 'en-CA';
-  const navigation = useNavigation();
-  const onDetails = useCallback(
-    ({id, exposureType}) => navigation.navigate('RecentExposureScreen', {id, exposureType}),
-    [navigation],
-  );
-
-  return (
-    <>
-      {exposureHistoryData.map((item, index) => {
-        return (
-          <Box key={item.timestamp}>
-            <Box backgroundColor="gray5" style={styles.radius}>
-              <Box paddingHorizontal="m" style={[exposureHistoryData.length !== index + 1 && styles.bottomBorder]}>
-                <TouchableOpacity
-                  style={styles.chevronIcon}
-                  onPress={() => {
-                    onDetails({id: `${item.id}-${item.timestamp}`, exposureType: item.type});
-                  }}
-                >
-                  <Box paddingVertical="m" style={styles.exposureList}>
-                    <Box style={styles.typeIconBox}>
-                      <Icon size={20} name={item.type === 'proximity' ? 'exposure-proximity' : 'exposure-outbreak'} />
-                    </Box>
-                    <Box style={styles.boxFlex}>
-                      <Text fontWeight="bold">{formatExposedDate(new Date(item.timestamp), dateLocale)}</Text>
-                      <Text>{item.subtitle}</Text>
-                    </Box>
-                    <Box style={styles.chevronIconBox}>
-                      <Icon size={30} name="icon-chevron" />
-                    </Box>
-                  </Box>
-                </TouchableOpacity>
-              </Box>
-            </Box>
-          </Box>
-        );
-      })}
-    </>
-  );
 };
 
 const NoExposureHistoryScreen = () => {
@@ -164,7 +121,6 @@ export const ExposureHistoryScreen = () => {
             <>
               <Box paddingHorizontal="xxs" marginLeft="m" marginRight="m" paddingBottom="m">
                 <ExposureList exposureHistoryData={mergedArray} />
-
                 <Box marginTop="m">
                   <Button variant="opaqueGrey" text="Delete All" onPress={deleteAllPlaces} />
                 </Box>
@@ -178,45 +134,11 @@ export const ExposureHistoryScreen = () => {
 };
 
 const styles = StyleSheet.create({
-  boxStyle: {
-    flex: 1,
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  bottomBorder: {
-    borderBottomColor: '#8a8a8a',
-    borderBottomWidth: 1,
-  },
-  textBox: {
-    flex: 1,
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-  },
-  radius: {
-    borderRadius: 10,
-  },
   noExposureHistoryScreen: {
     flex: 1,
     alignItems: 'center',
   },
   flex: {
     flex: 1,
-  },
-  exposureList: {
-    flexDirection: 'row',
-  },
-  boxFlex: {
-    flex: 4,
-  },
-  chevronIcon: {
-    alignItems: 'flex-end',
-  },
-  chevronIconBox: {
-    flex: 1,
-    justifyContent: 'center',
-  },
-  typeIconBox: {
-    flex: 1,
-    justifyContent: 'center',
   },
 });
